First Hughes Changes -- runnable build
Here is the first Hughes commit after cloning from Yocto. This
gives a relatively clean build. When loaded into a board, along
with a kernel and application, it actually does something. There
are still issues -- ping is missing from busybox, and the machine
type is V5, which will have to go to A9.

But basically, this gives a root file system that runs in a Callisto
terminal (which exists today), and is very similar to what will run
in the future Thebe terminal.